backend/requirements.txt
# Web Framework
fastapi==0.95.2
uvicorn[standard]==0.22.0

# Database
sqlalchemy==2.0.15
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
cryptography==41.0.5
google-auth==2.23.4
google-auth-oauthlib==1.1.0

# Data Validation & Parsing
pydantic==1.10.7
email-validator==2.1.0
python-multipart==0.0.6
orjson==3.9.10
ujson==5.8.0

# HTTP & Networking
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.1
certifi==2023.11.17

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
python-dotenv==1.0.0
python-json-logger==2.0.7 